refactor(product): split ToastBody under the 400-line shape cap
Extract expansion logic, details strip, excerpt, and copy controls into
ToastExpansion.tsx (242 lines). ToastBody.tsx now 375 lines, down from 490.

Pure move-and-import refactor — zero behavior change. All exports stable
(StatusToastBody, AnnouncementToastBody), so consumers and tests unchanged.

This unbreaks the repo-shape gate red on main since #1686.
